st-hal: Rework audio concurrency handling
Dedicated path: One change that affects dedicated path
usecases is to handle audio capture concurrency changes on
device events instead of stream events.

Non-dedicated path: The above change is also applicable. Also,
AR will transition to NLPI in order to avoid LPI to NLPI
transitions during audio caputre concurrencies. The reason for
this is that there could be a mismatch between the AFE channel
count and channel mask if transitions are allowed. The issue
comes from not being allowed to disable the backend device
during capture concurrencies, so this is avoided with this
change.

Change-Id: I82248dc506a79227d3d9f92064e560af350c590e